My Home: Refactor Launch Site card.
