+++ client/my-sites/customer-home/main.jsx
@@ -33,11 +33,13 @@ import { localizeUrl } from 'lib/i18n-utils';
 import { launchSiteOrRedirectToLaunchSignupFlow } from 'state/sites/launch/actions';
 import { bumpStat, composeAnalytics, recordTracksEvent } from 'state/analytics/actions';
 import isAtomicSite from 'state/selectors/is-site-automated-transfer';
+import { expandMySitesSidebarSection as expandSection } from 'state/my-sites/sidebar/actions';
 import StatsBanners from 'my-sites/stats/stats-banners';
 import isUnlaunchedSite from 'state/selectors/is-unlaunched-site';
 import { getCurrentUser, isCurrentUserEmailVerified } from 'state/current-user/selectors';
 import GoMobile from 'my-sites/customer-home/cards/features/go-mobile';
 import GrowEarn from 'my-sites/customer-home/cards/features/grow-earn';
+import LaunchSite from 'my-sites/customer-home/cards/features/launch-site';
 import Stats from 'my-sites/customer-home/cards/features/stats';
 import FreePhotoLibrary from 'my-sites/customer-home/cards/education/free-photo-library';
 import { getSelectedEditor } from 'state/selectors/get-selected-editor';
@@ -75,13 +77,6 @@ class Home extends Component {
 		},
 		trackAction: PropTypes.func.isRequired,
 		isStaticHomePage: PropTypes.bool.isRequired,
-	};
-
-	onLaunchBannerClick = e => {
-		const { siteId } = this.props;
-		e.preventDefault();
-
-		this.props.launchSiteOrRedirectToLaunchSignupFlow( siteId );
 	};
 
 	renderCustomerHomeHeader() {
@@ -157,7 +152,6 @@ class Home extends Component {
 
 	renderCustomerHome = () => {
 		const {
-			isAtomic,
 			isChecklistComplete,
 			needsEmailVerification,
 			translate,
@@ -171,8 +165,6 @@ class Home extends Component {
 			return <div className="customer-home__loading-placeholder"></div>;
 		}
 
-		const isPrimary = ! isAtomic && isChecklistComplete;
-
 		return (
 			<div className="customer-home__layout">
 				<div className="customer-home__layout-col customer-home__layout-col-left">
@@ -180,20 +172,7 @@ class Home extends Component {
 				</div>
 				<div className="customer-home__layout-col customer-home__layout-col-right">
 					{ siteIsUnlaunched && ! needsEmailVerification && (
-						<Card className="customer-home__launch-button">
-							<CardHeading>{ translate( 'Site Privacy' ) }</CardHeading>
-							<h6 className="customer-home__card-subheader">
-								{ translate( 'Your site is private' ) }
-							</h6>
-							<p>
-								{ translate(
-									'Only you and those you invite can view your site. Launch your site to make it visible to the public.'
-								) }
-							</p>
-							<Button primary={ isPrimary } onClick={ this.onLaunchBannerClick }>
-								{ translate( 'Launch my site' ) }
-							</Button>
-						</Card>
+						<LaunchSite />
 					) }
 					{ ! siteIsUnlaunched && <Stats /> }
 					{ <FreePhotoLibrary /> }
@@ -236,7 +215,6 @@ const connectHome = connect(
 		const siteId = getSelectedSiteId( state );
 		const siteChecklist = getSiteChecklist( state, siteId );
 		const hasChecklistData = null !== siteChecklist && Array.isArray( siteChecklist.tasks );
-		const isAtomic = isAtomicSite( state, siteId );
 		const isChecklistComplete = isSiteChecklistComplete( state, siteId );
 		const createdAt = getSiteOption( state, siteId, 'created_at' );
 		const user = getCurrentUser( state );
@@ -249,7 +227,6 @@ const connectHome = connect(
 			canUserUseCustomerHome: canCurrentUserUseCustomerHome( state, siteId ),
 			hasChecklistData,
 			isChecklistComplete,
-			isAtomic,
 			needsEmailVerification: ! isCurrentUserEmailVerified( state ),
 			isStaticHomePage:
 				! isClassicEditor && 'page' === getSiteOption( state, siteId, 'show_on_front' ),
