+++ client/my-sites/customer-home/style.scss
@@ -74,13 +74,6 @@
 			}
 		}
 	}
-	&__launch-button {
-		& .button {
-			font-size: 18px;
-			width: 100%;
-			height: 50px;
-		}
-	}
 
 	&__card-col {
 		display: block;
